Separate index.tsx to indicator/tab/tabs component
