+++ src/v2/constants.ts
@@ -1,0 +1,3 @@
+import { Dimensions } from "react-native";
+
+export const WIDTH = Dimensions.get('window').width;
